app/requirements.txt
flask
requests
python-dotenv
dictor